requirements.txt
python-keycloak
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.2
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.3
passlib==1.7.1
PyJWT==1.7.1
pytest==4.4.1
